requirements.txt
streamlit==1.50.0
spotipy==2.25.1
pandas==2.3.3
matplotlib==3.10.7
plotly==6.3.1
scikit-learn==1.7.2
python-dotenv==1.1.1
numpy==2.3.4
requests==2.32.5
